Make REBNOD a structure for clarity, malloc() release build
The REBNOD data type was somewhat unclear, making heavy use of void
pointers.  It is significantly more clear when seen as a structure
containing a pointer to another instance of itself (that is only good
when it is free.)

Adds some commentary regarding REBSER, and calloc()=>malloc() in release
build of the internal allocator.