Replace ad-hoc type checking from ADAPT
ADAPT was written in a way such that it would ensure you hadn't mutated
the parameters to be invalid types for the function about to be run.
It did so through with an ad-hoc type checker.  This gets rid of that,
switching to use the same type checking that dispatches functions in
the first place.