Put header bits into Reb_Node, kill small string pool
In R3-Alpha, there was a "small string pool" which allowed for
pooled allocations of size 8 or less.  This was possible because a
node could be as small as a single pointer--needing only that size
in order to track the next free pointer in the freelist.

There are some technical problems with the contract of using that few
bits for a node.  The signal of whether a node is used or not had to
live in the remaining bits--for those node types that did GC (the small
pool could not).  But there was a question as to what a pool cell that
had never been used would look like.  This meant it was all cleared to
zero and the client chose, but that has some undefined behavior issues
because the reads were done through different types than the writes.

Not having some kind of "shared header" also led to missing
opportuniites by which nodes of different types could signal what they
were if they used a common header format.  Then a pointer to a REBSER
or a REBVAL--for example--could be "sniffed" to tell which it was.

While it may seem like a loss to not be able to make these small
allocations, it actually isn't--because all those pooled allocations
also needed a REBSER node to point to them.  Given that fact, the
better option is to have a way for small things to live in the REBSER
node itself, in place of the information that would be used to track
the allocation.  This feature is partially implemented, and this is
a step further toward it.