Warn instead of crash on no top-level handler
There are some scenarios where you want to get debug output through
printing, e.g. with PROBE, and there's no top level handler in effect.
The issue with Ctrl-C during output needs to be addressed in a more
general way, but this converts the crashing assert into a message so
that you can take your chances if just looking to get debug output.