Strengthen "no leaks" assert on shutdown
The idea of tight control over the manually allocated series in the
system has been something moving incrementally forward over time since
Ren-C began tracking the distinction from R3-Alpha.  The manual leaks
are checked for at the end of every Do_Core, so there shouldn't be any
except for things created manually during startup that leak.

This takes care of the data stack, which appears to be the only thing
that wasn't being freed.  It panics on any leaked series at shutdown.