Add support for "Pairings", arbitrary GC roots
This adds support for the idea of arbitrary REBSER nodes serving as GC
roots.  At the moment the roots live in the same pool as the other
REBSERs, which leads to the somewhat undesirable "two pass" nature of
the GC...visiting all nodes first to find roots, and then later for
the sweeping phase.  However, review of GCs and the problems with the
current one suggests that a multiple pass or "copying"/"incremental" GC
which does more passes and is "slower but butter" is likely the best
option anyway, as the idea would be to not "stop the world".

It also adds support for a new "Pairing" concept, which is two REBVALs
that fit inside one REBSER node.  These are being tested with Ren-Cpp
as the API handles, acting as GC roots with the second REBVAL being
"meta" information about the value.