Fix help for undocumented functions, MAYBE
This fixes some problems that had gotten into the HELP of bare functions
that had no meta information, e.g. help on `f: function [x] [x]`.

It also changes the older IS experiment to be MAYBE, as a parallel to
ENSURE, and changes ENSURE to use MAYBE in its implementation.