Eliminate most `#pragma pack` specifications
Splitting one piece of the payload out as the "extra" leads to correct
structure packing without the need of a compiler-specific pragma.

Being able to leave it out carries the benefit of being able to leave
out long-winded explanations of why it is there, as well as not using
a non-standard compiler feature.