Reorganize Scan_Block, rename for consistency
This does a few surface level improvements of the source for Scan_Block,
including changing it to "Scan_Array" so that it's clear that it's
for reading paths, blocks, and groups.  Also renames "Scan_Source" to
"Scan_UTF8_Managed" to be more informative that the input needs to
be UTF8, and that the resulting array will be pre-managed by GC.

Also uses inline initializations in scan error, and moves the skip to
byte routine to the C defs so it can be inlined.

Includes a missing IF=>MAYBE fix.