Very basic prettiness fixes to HELP
As a start on not neglecting HELP, this does a few extremely basic
changes...like using spaces instead of tabs (tabs were 8 spaces and
wasting screen real estate, also they are tabs).  It also puts the
notes for a parameter on its own line, and moves the type notes onto
the same line as the parameter, in brackets and not parentheses.