Patch for chained functions on stack after R_REDO_CHECKED
The data stack is used to store the list of refinements, as well as
changing those refinement words into pointer entities to locate out
of order refinements being used.  But after the parameters are checked,
it's used to hold functions that are chained.

However, it's necessary for adapted functions to recheck the arguments
after running their injected code.  They do this by jumping up to the
argument loop.  There are no refinements to process at this point--so
the presence of the functions should just be ignored.