Simulate implicit termination for all arrays
R3-Alpha required all arrays to have a full REBVAL-sized "END!" cell
in order to indicate termination.  Hence instead of paying for
N x sizeof(REBVAL) for an array of length N, it was instead
(N + 1) x sizeof(REBVAL).  Ren-C sought to help reduce this overhead,
which is significant for large numbers of small arrays...by allowing
any pointer-sized value with a low bit of 0 to signal a terminator.

Though this has been tested on length 1 "singular" arrays that fit in
REBSER nodes, applying the method to dynamic arrays (or arrays that can
just resize from singular to dynamic) required changes so that it was
not assumed that a full cell's width of data is present in the last
cell of an allocation's "rest".  This preloads that spot with an
END marker that is unwritable in the debug build.

This also unifies termination techniques in general, to try and reduce
the chance for errors--including singling out those cases of setting
lengths that deliberately don't terminate for performance reasons,
via a separately named routine.