Remove seemingly unused code from scheme handlers
There was code for handling "port action maps", which seemed to be an
alternative for the switch statement in native scheme handlers.  A
comment said:

    static const PORT_ACTION File_Actions[] = {
        SYM_OPEN,     P_open,
        SYM_CLOSE,    P_close,
        0, 0
    }

    Register_Scheme(SYM_FILE, &File_Actions[0], 0);

But no instances of these existed, they all took a "PAF" (Port Action
Function) instead of one of these maps.

Deleted in attempt to further pare down the legacy code.