Fix TUPLE! crash after payload switch, round money
This addresses two test failures that had crept in after the payload
fix that were low on the priority list (MONEY! and TUPLE!).

Tuples can now be a maximum size of 7 elements instead of 10.  Hardcoded
numbers of "10" had to be adjusted.

Rounding money did some interpretation of the ROUND/TO amount, so that
the type of the /TO argument would influence the type of the result.
So rounding $10.51 to 0.1 would give you 10.5, for instance.  Not clear
quite how the code was supposed to work before, but this makes it
seem to work and is clearer.