Return PARSE to TRUE/FALSE, elaborate /? names
An experiment was run for a while which would return the input from
PARSE if it matched (instead of just TRUE) and then blank if it did
not match (instead of FALSE).  Then a PARSE? was offered that would
do only TRUE and FALSE results, limiting the return results to that
if someone tried to use a RETURN instruction inside the PARSE.

This was tried before CHAIN and the other function composers, so it
resulted in a duplication of PARSE's interface.  The idea had a bit of
an ergonomic problem in that it made the results of PARSE harder to
understand than TRUE/FALSE in the common case.  Given COLLECT as a
future possible direction, it may also confuse matters further.
Additionally, the usefulness of a return result the caller already has
is likely a poor choice...better would be something that parse knows
(e.g. the position of a failure).

This uses CHAIN to make a checking version to enforce the TRUE/FALSE
result, and restores the common baseline of TRUE and FALSE until more
thinking has been put into it.  It also removes /ALL from the non
compatibility PARSE.

This also takes the other /? refinements for functions and gives them
more descriptive names.