Fix stuck bind table entries on func spec failure
Though a "binding table" doesn't exist anymore, an analogous thing
exists to allow scribbling properties onto canon symbols (which would
need to be done via atomics in a threaded system).  However, all
annotated words must be un-annotated before a fail() happens -or-
the BUF_COLLECT buffer must contain entries of what to clean up.

Processing function specs in Make_Paramlist_Managed merged together
into a single phase for processing strings/types/words, but reused
the binding table mechanic to check for duplicates.  In that process
it could fail mid-spec-analysis, leaving stray entries in the table.

This does the duplicate check at the end, when the only failure left
to give back is the failure of the duplication itself...and cleans up
before giving that error.