REDUCE+COMPOSE use safe enumeration, kill unused refinements
This updates REDUCE and COMPOSE to create a single frame for the course
of their evaluations, instead of starting one and tearing it down again
each time.  In the process they also lock the input array to protect
against crashes for the duration of their operation.

This also gets rid of the /ONLY and /NO-SET options of REDUCE, which
were not used and not being tested.  If such features are needed, they
can be reconsidered in general with the idea of a REDUCE that might
need other "hooks" (something like a collecting REDUCE-EACH)