Fix VARARGS! GC bug via simplification
Whether a VARARGS! sources from an actual FRAME! (which can go off the
stack) or a BLOCK! serving as a parameter manufactured parameter pack
(which it protects), the source is able to chain to a "subfeed" of
values to exhaust before resuming from its own data.

The GC was attempting to protect the subfeed in the FRAME! case, when
the frame already protects it.  It wasn't getting the logic right, but
rather than improving the logic this takes it out--since there is no
need to doubly protect it.