Add /WHERE feature to FAIL
This lets the caller of FAIL specify a FRAME! or word of a frame where
an error should be reported.  With this, it's possible to indicate a
location other than the stack of the point of failure itself--which
may be more useful (e.g. to implicate an incorrect caller instead of
showing the guts of a library routine that is behaving as expected).

The notion of implicating an argument could--in debug modes--be able
to implicate the source expression from which the argument was
gathered.  That would require storing additional positioning
information about where each argument came from.