MAYBE as native, dialected ASSERT, VERIFY, more
This is a general review of ASSERT and related routines.  It starts with
the removal of the /TYPE refinement of ASSERT, instead leaning on the
more generic functionality of ENSURE.  ENSURE was already implemented
in terms of MAYBE, so this takes the ASSERT/TYPE code and updates it
to a more general "tester" that can accept functions as well.

Then ASSERT is distinguished from VERIFY by being commentary unless a
debug mode is enabled.  It is dialected to allow for TAG! and ISSUE!
to enable and disable tests.  Due to existing assumptions that assert
always runs, the debug mode is enabled by default.