Remove spurious assert about f->out being end ET_LOOKBACK
It is legal for f->out to be an END marker when Do_Core is dispatched
on ET_LOOKBACK, because the code for seeing if there's anything "back
there" was factored out into the "optimized" fetch/do.  So it does
not delegate those cases to ET_WORD any longer, and END markers can
thus happen.