Don't evaluate non-blocks passed to PRINT
The logic for PRINT was not noticing that things like WORD! should not
be evaluated when printed as a direct argument.  So it errored on:

    x: 'foo
    print x

This adds the necessary test for it being a block *and* not /QUOTEd
to ask for evaluation.

Also brings R3-LEGACY parse up to date with the reversion of its return
results back to TRUE and FALSE by default.