Fix REMOVE copying of implicit terminator, related cleanups
This is cleanups that were done while chasing an issue where REMOVE
was not using TERM_SERIES, but trying to bit-copy the terminator from
its existing location.  This is incompatible with implicit termination,
where there *might* be an incomplete bit pattern for a cell that passes
for an END marker, which should not be moved to other cell locations.

Finding it required factoring the Do_Signals mechanic in the evaluator
loop so it ran at both the lookahead-fetched and top-of-switch points.
(Previously only ran at the top of switch, e.g. the do-next: label.)