Fix "unwritable in rest" assertion, clarify C++-only asserts
A C++-only check was being run in the C build.  This solves the issue,
as well as #ifdef'ing out the checked flag unless the build is C++.

(A source of confusion is that a flag that was once checked in all
debug builds was moved to the C++ build only, due to feeling that the
manual need to "VAL_INIT_WRITABLE_DEBUG()" on every REBVAL declaration
was "cluttering up the code".  The change which moved it to C++-only
didn't re-evaluate all the terminology used to make it clear, so this
tries to do that re-evaluation.)