Narrowly disable underlying match assert for hijack proxies
To speed function dispatch, the "underlying" function of a chain of
compositions (adaptations, chains, specializations) is stored in the
REBSER of a function itself.  This means that a specialization of a
specialization of a chain of an adapted specialization (for instance)
does not need to dig through all those compositions at call time to
figure out what exact set of variables need to be in the frame that
the whole composition needs to use.

The debug build asserts that the cached value is the same as what you
would have gotten by following all the links and looking manually.  In
the case of proxying functions returned from hijacking, however, they
will *not* match.  This is on purpose.

Rather than disable the assert entirely, this sets a flag on functions
that are proxies to only disable checking in this case.  Because series
flags are at a premium, the flag is placed on the canon function value
in the [0] slot of the paramlist of the proxied function.