Move `func [[catch] x][...]` support to <r3-legacy>
Rebol2 had "catch" and "throw" attributes that could be placed on
functions by putting them in a block at the head of a function
definition.  R3-Alpha silently ignored these.  Ren-C did as well, but
that ignoring became broken.

Rather than fix the C code that ignores it inside function generation,
this moves support for ignoring it to <r3-legacy>.