Undo namings /? => BRANCHED?, MATCHED?, RAN?...
Concern over the "illiterate" symbolic nature of having a refinement
called /?, and figuring that it would always be run from a
specialization anyway, led to an attempt to give the /? more descriptive
names for each function it was applied to.

This inspired debate from @ingo over the name choices, as seen from the
HELP.  The HELP visibility was one point not considered, that it would
be "clutter" and draw attention to questions of the choices of names.

Reconsidering it under this light, the simplicity of just calling it
/? seemed more favorable.