Add "ENFIX", and "enfixed" DEFAULT
Because there is only one kind of FUNCTION! in Ren-C, there is no OP!.
Hence you can't say `+: make-op :add` (or equivalent) because the idea
of "infixness" comes from a bit set in the binding.  Up until now this
means you have had to write `set/lookback '+ :add` or a wrapper like
`set-infix '+ :add` (which checks for arity 2 specifically).  This is
because there is no way to "decorate" a function value to say that it
is infix, in a way that wouldn't be lost by the assignment.

But there is also the feature for quoting SET-WORD! or SET-PATH! as the
first argument of a prefix/infix/n-fix function.  This adjusts that
feature such that it will trust the function's assignment and not
overwrite it, but instead convert it into a GET operation...so whatever
value the SET-WORD! or SET-PATH! has at the end of the operation is
considered to be the result of the assignment.

This leverages that ability to make ENFIX, so:

    +: enfix :add

The enfix operation quotes the +: on the left, and then has the
responsibility of setting it.  It does so with a SET/LOOKBACK.  Then,
in order to enforce consistency, after enfix is done running the +:
evaluation fetches the value of +.  (This is to make sure that despite
the delegation of the assignment, chains such as `x/y: z: foo: blah`
will get the same values for x/y, z, and foo regardless of the return
result of BLAH.)

Additionally this makes DEFAULT use the same mechanism.  It works for
SET-WORD!s, or SET-PATH!s...and will overwrite any BLANK! or not set
values in those words and paths:

    >> o: object [x: 10 y: _]

    >> o/x: default 20

    >> o/x
    == 10

    >> o/y: default 20
    == 20

    >> o/y
    == 20

Only paths without GROUP!s in them are supported for now, to avoid a
double evaluation of that group.