Move several functions to headers so they can inline
Although in today's C and C++ "inline" is just a suggestion (and the
compiler will do whatever it feels like, basically removing inline
from the consideration)...having the code available in headers gives
more options for the compiler to avoid things like pushing and popping
stack space on every call.

There was a measurable amount of overhead on operations that inlining
could be eliding.  R3-Alpha addressed this on a much smaller scale
with macros, but this pushes toward taking much more advantage of the
inlining abilities of the compiler--with performance gains (20% on
basic idle dispatching, e.g. `loop 1000000 [x: 1 + 2]`)

Also aliases struct Reb_Frame to REBFRM, a concept initially treated
with wariness but now an obvious win.