Put type bits in top positions for shifting out
The 6-bit type (or KIND) used to live in the low 8 bits of a value
header, and needed to be masked and shifted right 2.  This changes the
strategy so that the bits are in the high positions, and only need to
be shifted right (with zeros filling from the left).

While performance-wise this doesn't appear to make a ton of difference,
it does simplify things...because previously the types were staggered
in order to allow them to be checked with a masking operation only.
Now they can be checked against 0-based constants with a shifting
operation only.  This eliminates ideas like REB_MAX_0 and the other
XXX_0 based transformations...since the 64 types are numbered 0
up to 63.