Fixes for BLANK! and empty block in PARSE handling, tests
When VAL_TYPE() stopped working on END markers, this was causing some
crashes in PARSE.  The reason was that it would find itself at the end
of a block, holding an end marker in a value...and then try testing
it...expecting a failure (as R3-Alpha was laissez-faire about tests
on END, if you said IS_INTEGER(value) and value was an end marker, it
would just say false).

A patch to keep PARSE from crashing did not accomodate cases where it
really would be okay to run a rule even if one were at an END--namely
if the rule would be able to accept no input.  This applies to blanks
(formerly NONE!) and block rules which bottom out at empty blocks or
at blanks.

This also breaks PARSE tests out into their own file, as a start on
the idea of breaking tests into more files in general.