INIT_CELL_WRITABLE on eval cell before arg fulfillment
The bits inside the cell-sized slot of a frame are fair game for any
non-cell purpose unless tasked otherwise.  The check that ensures
cells have been pre-formatted in the C++ debug build for writing was
being tripped by cases from an APPLY entry point that weren't formatting
it before jumping to argument fulfillment.

This throws in a format of the eval cell before all argument fulfillment
cases in the debug build.