Separate DUMP routine (like ??) from PROBE
An experiment of trying to get rid of ?? and add more functionality to
PROBE was probably not a great idea.  This undoes that and calls the
dialected routine DUMP.  It is set to be ?? in R3-Legacy.

Also reverts an R3-Legacy bad use of ENFIX, which needed to be in
a GROUP! to be added to the context...but ENFIX inside a group can't
work, because then it can't quote backwards to get at the SET-WORD!.