Manage array when appending to self
The case of:

    block: []
    append block block

Avoids problems by copying the underlying array before starting the
append.  This array was not being managed or freed, so it was asserting.

Given that it's uncommon it's not clear whether the overhead of tracking
whether a copy was made or not and then manually freeing it is a win
in the long run vs. just managing it if the case comes up.  This
manages for now just to stop the assert.