Set-Words/paths pushed, lookahead in lookback operations
(Note: r3-legacy compatibility switch as `no-infix-lookahead`)

Typically the evaluator is "greedy", so here it sees 10, then sees an
= operation whose first argument gets 10, and then as much is consumed
as possible...so `add 5 5` is processed:

    >> 10 = add 5 5
    == true

This greediness leads to errors if one doesn't see that an expression
is actually incomplete, e.g. before doing an equality test:

    >> add 5 5 = 10
    ** Script error: expected logic! not integer!`

There it's the case that for the second argument to add, 5 keeps looking
and turns into the logic expression (5 = 10) as the second argument.

It's a simple and longstanding rule.  But R3-Alpha had a strange infix
"feature" that twisted it the other way:

    >> 5 + 5 = 10
    true

    >> 10 = 5 + 5
    ** Script error: expected logic! not integer!

After long wishing to make this behavior go away unless there was a
*really* good reason for it, this commit finally excises the idea.  The
straw breaking the camel's back in this case was what it did to
an attempt to shift SET-WORD! and SET-PATH! to be processed without
making a new frame.  Instead they would be pushed to a pending state
on the data stack, and flushed at a convenient time.

This showed these troubling variations in R3-Alpha's behavior:

    >> 10 = probe 5 + 5
    == true

    >> 10 = x: 5 + 5
    == true

The addition or removal of a PROBE or SET-WORD! doesn't seem like it
should transform what was previously an erroring case into a correct
one.  Or vice-versa; removing the probe and the x: from the above code
does not seem like it should cause an error.

Despite the compatibility issues, the balance is in favor of fixing the
behavior.  Parentheses can be used to stylize code in such a way that
it would work before or after the rule change, which is at least some
benefit to those trying to write code that would work under either rule.