Split the CORE-TESTS.R file.
The core-tests.r file is a monolithic test file that appears to have been amalgamated from separate files, each of which is represented by a comment. This commit splits core-tests.r into separate files again, but using a shallower organisation than that described by the comments. These comments will be retained in the separate tests files for now.

A core-tests.SPLIT.LOG file is used to document in this commit how the split of files was made. It will be subsequently deleted.