Hacky hash to allow e.g. unique [[1 2] [1 2]]
This adds support for a very poor hash of blocks (their length) as
a spur-of-the-moment way to allow them to be used with the set
operations.  Coming up with a better hashing strategy is left as an
exercise to those interested.

This also adds support for using locked ANY-ARRAY! types as keys in
MAP!, which is probably the longer term strategy for series keys in
the immutable-source-by-default future.