checked RETURN: types in function specs, notes in HELP
This adds the long-planned feature of being able to document and check
the return types of functions.  It uses the planned notation from
R3-Alpha and Red of a RETURN: set-word in the function spec, followed
by type and description as with other parameters.

***NOTICE*** This also implements an experimental-yet-important Ren-C
"rule" that FUNC and FUNCTION must return a value unless they explicitly
say they do not.  This means that return values--like parameters--
default to not being optional.  The PROC and PROCEDURE generators are
the most likely intent behind functions that want to return voids, as
opposed to "sometimes void, sometimes not".  They also have the
advantage of suppressing values to prevent them from accidentally
"falling out" the bottom of the function body...which can be confusing
to the caller (if meaningless) or even a potential security issue.

This also implements another rule, which is that FUNCTION! is not
considered by default to be a legal parameter for function types--nor
a return type from them.  This is because like voids (already disabled
by default in R3-Alpha for function arguments), FUNCTION! values must
be handled with particular care and awareness that they are function
values.  The best way of documenting this fact is to annotate the
type to show you know that.  Similar rules apply for return types, in
that callees must be particularly prepared to handle the behaviors of
a function if that's what they receive.

Given that the generators can be adjusted to one's tastes (e.g. making
one's own FUNCTION, FUNC, PROCEDURE, PROC) then these policies are
not hard and fast rules for all coders.  But after absorbing the
initial conversion, they are likely to improve the debuggability and
coherence of the code.