Prevent failed SET in PARSE from overwriting with BLANK!
Existing rule for parse was such that a failed SET would set the
quantity to BLANK!, as opposed to leave it as-is.

In the general spirit of "rollback", it is best if you can write
something like `(data: "default") opt set data: word!` and not have
the failed optional rule affect state.

This is also what Red does in the same situation.