DEFAULT evaluates blocks
DEFAULT is still a work in progress.  But this makes it evaluate
blocks... a question being that an /ONLY version can't be implemented
because you can't do infix dispatch off of paths.

Relocates the ENFIX and DEFAULT functions to base-funcs from base-infix
due to upcoming dependency in FUNCTION, which is used before the
infix definitions get run (as well as in the infix definitions).