<in>, <with>, <static> function specs...plus defaults
This moves the /WITH refinement on FUNCTION as a feature to be the
<in> tag, and switches the block-based statics aspect of that /WITH
to be a <static> feature.  So instead of:

    f: function/with [spec...] [body...] [x: 1 y: 2]

You can write:

    f: function [spec... <static> x (1) y (2)] [body...]

The /EXTERN feature on the other hand took the "with" name, as a way
of defining the outer words you want to define as "being brought in
with" the function defintion.  So instead of:

    f: function/extern [spec...] [body...] [a b c]

You can write:

    f: function [spec... <with> a b c] [body...]

Along with these conveniences is the ability to use parentheses as a
way of providing defaults to refinement args, that will become their
value if they are not passed in (or passed in and BLANK!).  Whether
BLANK! should be overridden or not is a question up for debate.

All parenthesized expressions are invoked in the context of function
generation--not the function body--which means it only runs once
(hence good for reducing total computation if the model fits) as well
as bridging data in variables the function body won't have access to
when it runs otherwise.