Add support for ELSE infix function generator
This commit does a little bit of tidying to the method for dispatching
to Rebol functions from a variadic C call (similar to printf for passing
multiple args to a function).  As that feature has been enhanced, it
has become possible to have a sort of "downsizing" response to handlers
that accept lower arity than the number of potential parameters passed,
so the code calling it can be less conditional (and less errant, as
the detection mechanisms for detecting call validity are tricky...and
variadics/etc. would throw a wrench into simple methods).

But it most notably adds support for ELSE, to permit the construction:

    IF condition [foo] ELSE [bar]

The formulation of it does not break rules of Rebol design (e.g. by
employing variadics, or even quoting).  Instead it defines ELSE as
an infix function generator.  The resulting function takes a logic.

    >> which: "apple" else "banana"

    >> which true
    "apple"

    >> which false
    "banana"

Each call to ELSE creates a specialization of EITHER.  Consequently,
it has the property of running code if you pass it a block:

    >> runner: [print "truth"] else [print "falsehood"]

    >> runner true
    truth

    >> runner false
    falsehood

The commit also adds a feature to IF (when /ONLY is not used) so that
in addition to having special "running" behavior of blocks, it will also
run arity-0 and arity-1 functions.  If the function is arity-0, it will
only be run if the condition was true.  If the function is arity-1, it
will be run if the condition was true or false, with the logic passed.

There's not heavy overhead to the dispatch to the function, relative
to using blocks.  But the creation of a function and GC'ing of the
function on each time does have overhead; costing about 4-5x as much
as using an EITHER in practice.  That means this is more of a "thought
experiment" and test of the evaluator (customized infix, specialization)
than something one would want to use.

That said, caching techniques for function generation is probably a
generally good idea...where successive calls to the same function
generator with the same data would not make a new function identity
each time.  If that were to happen, the difference in performance might
become basically negligible.