Check one signal variable per loop instead of 2
R3-Alpha's evaluator loop had a countdown (Eval_Count) which was
decremented on every step.  When this counter reached zero, it would
call Do_Signals to process any "signals"...which could be requests for
garbage collection, network-related, Ctrl-C being hit, etc.

But in addition, on every step it would also check the Eval_Signals mask
to see if it was non-zero.  If it was, then it would call Do_Signals,
regardless of the Eval_Count.

While a broader review of how signals would work in Ren-C is pending, it
seems best to avoid checking two things instead of one to dispatch to
the same routine.  So only the Eval_Count is checked.  Places that set
Eval_Signals also reset the Eval_Count to 1, and if the signals are
not fully cleared by Do_Signals then it is reset to 1 instead of the
full countdown EVAL_DOSE.