Eliminate Reb_Eval_Type in favor of reordering Reb_Kind
There are several demands on the orderings of the Reb_Kind enumeration.
R3-Alpha had the idea of using a table to quickly reduce the enumeration
into values that needed evaluator behavior, and those that did not.
So there was an ET_SET_WORD mapped to by REB_SET_WORD, and an ET_SELF
for several types that just evaluated to themselves.

Ren-C in Coherence One got rid of this table and just switched on
REB_XXX types.  Later, the convenience of being able to know if a type
had evaluator behavior or not--in order to do certain optimizations--
led to the table being added back again.

Since that time, still more changes have happened that lead it to be
feasible and interesting to rethink the ordering of the types in the
Reb_Kind table...in such a way that they group by evaluating types
at the lower numbers.  This does a bit of shuffling but manages to
preserve some attempted optimizations--in particular the use of a
"fake" boolean masquerading as the type kept by the evaluator, to
encode the lookback state as well as set the eval_type for the
next operation.

Overall it simplifies by getting rid of Eval_Table and ET_XXX, though
there are more improvements that could be done.