Move EVAL body into dispatcher (native)
Though not a "normal" native, it's now more feasible to move EVAL
cleanly into a native-dispatcher-like-thing.  This is actually a bit
more like the original reevaluation mechanism in DO, just interacting
with the frame state instead of putting a bit on the value.

Because it evaluates the argument into a stack cell instead of the
frame directly, this means EVALs themselves get a little slower.  But
it reduces the overhead on every other non-eval call, which is overall
a benefit.  (It also means there can be less code and more focus on
speedups that benefit other functions.)