Performance Improvement: Don't double walk arguments
R3-Alpha did not have the concept of access to the iteration state of
a frame by the GC, e.g. knowing how far it had gotten in the parameter
list.  As a result, the arguments to a function had to be pre-filled
with something that wasn't garbage.  This effectively produces two
phases of walking the arguments list...once to make sure the GC won't
trip on it, and once to fill them.

Ren-C has f->param and f->arg to examine.  So this implements the idea
that the GC only goes up to the f->arg that lines up with f->param to
protect.  This required some refactoring, because filling the args
with specialization data was previously done in the "pre-walk", so
that had to be untangled so that the evaluator loop does not see
specialized values in the argument slots, but has access to them via
a separate variable.

Also includes an optimization by which the eval cell is used when a
function only has one argument, and so that the cell is available for
use by other functions that have more than one argument.