Centralize subfeed code to contain its "trick"
For purposes of garbage collector safety, if a frame is going to have
a variadic "feed" to which it is attached, it needs to be GC safe.
A trick is used to reuse the ->special field from specializations for
this purpose, even though it is not an array type...since REBARR and
REBVAL share a common header.

Something about that trick was not going right in optimized builds.
This does some centralization to try and make the (still largely
experimental/exploratory) VARARGS! code better handle the rule.

Future approaches will likely find a way to chain these inputs directly
into the ->pending list...making it a general macro facility.  Then
there would not be a subfeed, just ->pending...which could similarly
cue based on whether it was a single value, an END, or an array.  So
there would still be similar logic, just not using ->special.