Initialize frame single cell in DO FRAME!
There is a single cell's worth of storage in the frame.  It is needed
to hold the cell of an EVAL live while its D_OUT is being calculated,
but has been applied to the purpose of holding the argument to a
single-argument function (vs. pushing a chunk).

It was not being cleared with GC safe bits when a frame evaluation was
running.  This does that, and also applies it to SWITCH and CASE which
needed temporary GC safe cells, and were re-using arguments in a way
that makes them harder to explain (and also cost extra assignment to
put a LOGIC! variable into a local).