Return documentation for natives, checked in debug build
This moves return type checking (or turning into voids, as for
PROC) into a separate dispatcher.  This means checking to see if the
return type needs to be checked itself is not done for every function,
only for those requiring it in their own dispatcher.

It's also a first cut at adding return description support for natives.
Since there is no return type checking in each native's implementation,
this means they won't be checked in the release build.  However, the
descriptive information is still encoded in their meta-objects so that
the help shows a return value and description.

In order to facilitate the debug build checking the return values, the
natives are processed differently--preserving the RETURN: key in their
paramlist.  This means the debug build runs noticeably differently from
the release build, in the sense that the single-argument optimization
will be used more rarely (by 0-argument natives instead of 1-argument
ones).  The codepath is still exercised, however.

The first thing this addresses is the return types of natives that
give back voids or functions, as those must be explicit--following the
same rules as for func.  This is still a new idea, and may not be
permanent...but the return type annotations do no harm regardless.