Add BRANCHER native behind ELSE, support in CASE
This adds a somewhat faster native function generator for BRANCHER,
which means ELSE usage is less costly...though still significantly more
expensive than EITHER.  (Further optimizations are possible.)

In addition, it changes CASE so that like IF and UNLESS, it gives the
opportunity to failed branches to run if they are single-arity functions
that can take logic.  This also means that CASE will run 0-arity
functions by default...so the /ONLY option is added to suppress that
behavior.

Taking it a little bit further, it also adds brancher support to WHILE
and UNTIL.  This means a single arity function willing to take a LOGIC!
can be used to get the "last say" after the condition fails.  (It does
not currently run on BREAK, because breaks emerge from the body itself,
so it would be effectively getting notification that the body ended
the loop...which it shouldn't need since it was what ended it.)  This
is an experiment and is to be evaluated and possibly applied to other
loop functions.