Exercise "pairing" REBSERs in PAIR!, in BRANCHER
For several intended usages, a certain kind of REBSER which can contain
exactly two REBVALs was introduced.  It is intended to be the means by
which API handles to REBVALs are given out and managed, by allowing
a "meta" REBVAL to be used to track the intended lifetime of the
"in-band" REBVAL.

However, it is also intended to be the type behind MAP! key/value
storage...and generic PAIR!s in general.  This gives the type more
exercise by using it to back the PAIR! datatype, and to hold the two
branches in a BRANCHER.

This makes numeric pairs cost slightly more, though it points toward a
path for resolving the loss of precision or memory of whether integer
or floating point values were used to construct the pair.  It basically
does not make sense to have to document another numeric type and
range besides INTEGER! and DECIMAL!, because you cannot fit two of
those in a PAIR! as it is...so the calculations on coordinates are
lossy and complex.